+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Projects: Always read `scout.local.yaml` even if there is no `scout.yaml` file.
 - Projects: Always apply project level `filter` to scans (AND combine with scan filters).
+- Scan Results: Add `exclude_columns` parameter for reading parquet reuslts to optionally reduce memory usage.
+- Scan Results: Pre-fetch optimization for S3/remote parquet files.
 
 ## 0.4.10 (21 January 2026)